requirements.txt
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
python-dotenv==1.0.1
streamlit==1.30.0
beautifulsoup4==4.12.2
requests==2.31.0
google-generativeai
rank_bm25
SpeechRecognition
fastapi
uvicorn
